some advances

+++ PARTE2/main/commi2c.h
@@ -34,7 +34,7 @@ public:
 
 	void receiveHandler(int numBytes);
 
-	void msgDecoder(int last8, int first8);
+	void msgDecoder(byte last8, byte first8);
 
 	void readADC(int address);
